requirements.txt
fastapi
uvicorn
gunicorn
pandas
numpy
faiss-cpu
sentence-transformers
